# docs/serving/openapi.json
{
  "components": {
    "schemas": {
      "ForecastPointV1": {
        "description": "Point-only compatibility forecast for one hour.",
        "properties": {
          "load_mw": {
            "title": "Load Mw",
            "type": "number"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "timestamp",
          "load_mw"
        ],
        "title": "ForecastPointV1",
        "type": "object"
      },
      "ForecastPointV2": {
        "description": "Calibrated point and quantile forecast for one hour.",
        "properties": {
          "p10_mw": {
            "title": "P10 Mw",
            "type": "number"
          },
          "p50_mw": {
            "title": "P50 Mw",
            "type": "number"
          },
          "p90_mw": {
            "title": "P90 Mw",
            "type": "number"
          },
          "point_mw": {
            "title": "Point Mw",
            "type": "number"
          },
          "timestamp": {
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "timestamp",
          "p10_mw",
          "p50_mw",
          "p90_mw",
          "point_mw"
        ],
        "title": "ForecastPointV2",
        "type": "object"
      },
      "ForecastRequest": {
        "additionalProperties": false,
        "description": "Validated request contract shared by v1 and v2 forecasts.",
        "properties": {
          "history": {
            "description": "Sorted, unique, regular hourly positive load history.",
            "items": {
              "$ref": "#/components/schemas/HistoryRecord"
            },
            "maxItems": 1344,
            "minItems": 336,
            "title": "History",
            "type": "array"
          },
          "horizon_hours": {
            "default": 168,
            "description": "Number of hourly forecasts requested.",
            "maximum": 168.0,
            "minimum": 1.0,
            "title": "Horizon Hours",
            "type": "integer"
          }
        },
        "required": [
          "history"
        ],
        "title": "ForecastRequest",
        "type": "object"
      },
      "ForecastResponseV1": {
        "description": "Deprecated version 1 point-only serving response.",
        "properties": {
          "api_version": {
            "default": "v1",
            "title": "Api Version",
            "type": "string"
          },
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "forecasts": {
            "items": {
              "$ref": "#/components/schemas/ForecastPointV1"
            },
            "title": "Forecasts",
            "type": "array"
          },
          "horizon_hours": {
            "title": "Horizon Hours",
            "type": "integer"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "origin": {
            "format": "date-time",
            "title": "Origin",
            "type": "string"
          }
        },
        "required": [
          "model_version",
          "origin",
          "horizon_hours",
          "degraded",
          "forecasts"
        ],
        "title": "ForecastResponseV1",
        "type": "object"
      },
      "ForecastResponseV2": {
        "description": "Version 2 probabilistic serving response.",
        "properties": {
          "api_version": {
            "default": "v2",
            "title": "Api Version",
            "type": "string"
          },
          "degraded": {
            "title": "Degraded",
            "type": "boolean"
          },
          "fallback_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fallback Reason"
          },
          "forecasts": {
            "items": {
              "$ref": "#/components/schemas/ForecastPointV2"
            },
            "title": "Forecasts",
            "type": "array"
          },
          "horizon_hours": {
            "title": "Horizon Hours",
            "type": "integer"
          },
          "model_version": {
            "title": "Model Version",
            "type": "string"
          },
          "origin": {
            "format": "date-time",
            "title": "Origin",
            "type": "string"
          }
        },
        "required": [
          "model_version",
          "origin",
          "horizon_hours",
          "degraded",
          "fallback_reason",
          "forecasts"
        ],
        "title": "ForecastResponseV2",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "HistoryRecord": {
        "additionalProperties": false,
        "description": "One timestamped observed load provided to the serving endpoint.",
        "properties": {
          "load_mw": {
            "description": "Observed positive electrical load in MW.",
            "exclusiveMinimum": 0.0,
            "title": "Load Mw",
            "type": "number"
          },
          "timestamp": {
            "description": "Timezone-aware hourly observation timestamp.",
            "format": "date-time",
            "title": "Timestamp",
            "type": "string"
          }
        },
        "required": [
          "timestamp",
          "load_mw"
        ],
        "title": "HistoryRecord",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "GridCast Forecast Serving",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/health": {
      "get": {
        "description": "Expose non-sensitive process and serving state.",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Health Health Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Health"
      }
    },
    "/live": {
      "get": {
        "description": "Return liveness independent of model-load state.",
        "operationId": "live_live_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Live Live Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Live"
      }
    },
    "/metrics": {
      "get": {
        "description": "Return isolated Prometheus exposition text.",
        "operationId": "prometheus_metrics_metrics_get",
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Prometheus Metrics"
      }
    },
    "/ready": {
      "get": {
        "description": "Return readiness only after load and warm-up complete.",
        "operationId": "ready_ready_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Ready Ready Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Ready"
      }
    },
    "/v1/forecasts": {
      "post": {
        "description": "Serve deprecated point-only compatibility forecasts.",
        "operationId": "forecast_v1_v1_forecasts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponseV1"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forecast V1"
      }
    },
    "/v2/forecasts": {
      "post": {
        "description": "Serve calibrated point and interval forecasts.",
        "operationId": "forecast_v2_v2_forecasts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForecastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ForecastResponseV2"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Forecast V2"
      }
    }
  }
}